cloud-2022/function_app/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
beautifulsoup4
lxml
requests
psycopg2-binary